apps/api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
gunicorn>=21.0.0
redis>=5.0.0
prometheus-fastapi-instrumentator>=6.1.0
biopython>=1.81
xmltodict>=0.13.0
pandas>=2.0.0
requests>=2.31.0
regex>=2023.0.0
